third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-border-style-shorthand.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>

description("Tests that the border-style shorthand is computed properly.")

var testContainer = document.createElement("div");
testContainer.contentEditable = true;
document.body.appendChild(testContainer);

testContainer.innerHTML = '<div id="test">hello</div>';

e = document.getElementById('test');
computedStyle = window.getComputedStyle(e, null);

e.style.borderStyle = "solid dotted";
shouldBe("computedStyle.getPropertyValue('border-style')", "'solid dotted'");

e.style.borderStyle = "solid dotted groove dashed";
shouldBe("computedStyle.getPropertyValue('border-style')", "'solid dotted groove dashed'");

document.body.removeChild(testContainer);

</script>
</body>
</html>
